Now checking additional binaries before actually copying

+++ src/core/portablemc/__init__.py
@@ -791,17 +791,20 @@ class Start:
                                 shutil.copyfileobj(native_zip_file, native_file)
 
         for bin_file in self.bin_files:
-            bin_name = path.basename(bin_file)
-            # Here we try to remove the version numbers of .so files.
-            so_idx = bin_name.rfind(".so")
-            if so_idx >= 0:
-                bin_name = bin_name[:so_idx + len(".so")]
-            # Try to symlink the file in the bin dir, and fallback to simple copy.
-            bin_dst_file = path.join(bin_dir, bin_name)
-            try:
-                os.symlink(bin_file, bin_dst_file)
-            except OSError:
-                shutil.copyfile(bin_file, bin_dst_file)
+            # Resolve a potential symlink and check if the binary exists before linking.
+            bin_file = path.realpath(bin_file)
+            if path.isfile(bin_file):
+                bin_name = path.basename(bin_file)
+                # Here we try to remove the version numbers of .so files.
+                so_idx = bin_name.rfind(".so")
+                if so_idx >= 0:
+                    bin_name = bin_name[:so_idx + len(".so")]
+                # Try to symlink the file in the bin dir, and fallback to simple copy.
+                bin_dst_file = path.join(bin_dir, bin_name)
+                try:
+                    os.symlink(bin_file, bin_dst_file)
+                except OSError:
+                    shutil.copyfile(bin_file, bin_dst_file)
 
         self.args_replacements["natives_directory"] = bin_dir
